Added a new method getWidgetPosition()
This method will return the position of a widget within the column the
widget is currently in. This is NOT the same as the position "in the
grid" as it might be interesting to save the order of the widgets in the
grid. However, to get the column of the widget, use <widget>.column
instead.